Consolidate duplicated code
